Restructure project files; get sphinx to work properly
