+++ qucumber/__init__.py
@@ -1,4 +1,1 @@
-from .rbm import *  # noqa
-
-
 __version__ = "0.1a"
