--- qucumber/observables/__init__.py
@@ -1,1 +1,0 @@
-from .observables import *  # noqa
